+++ ui/restmodel/src/main/scala/pl/touk/nussknacker/restmodel/displayedgraph/DisplayableProcess.scala
@@ -47,7 +47,7 @@ import pl.touk.nussknacker.engine.graph.NodeDataCodec._
 @JsonCodec(decodeOnly = true)
 case class ProcessProperties(typeSpecificProperties: TypeSpecificData,
                              additionalFields: Option[ProcessAdditionalFields] = None,
-                             subprocessVersions: Map[String, VersionId] = Map.empty) {
+                             subprocessVersions: Map[String, Long] = Map.empty) { //TODO: field subprocessVersions is deprecate - to remove
 
   def toMetaData(id: String): MetaData = MetaData(
     id = id,