+++ docs/Changelog.md
@@ -10,6 +10,7 @@ Nussknacker versions
 
 1.3.0 (Not released yet)
 * [#2741](https://github.com/TouK/nussknacker/pull/2741) Remove custom scenario (custom process)
+* [#2773](https://github.com/TouK/nussknacker/pull/2773) Using VersionId / ProcessId / ProcessName instead of Long or String
 
 1.2.0 (Not released yet)
 ------------------------